+++ Changes.md
@@ -29,7 +29,7 @@
 * mailheader: fully quality header name in _remove_more #2647
 * haraka: Connection.createConnection is not a constructor #2816
 * problems with japanese characters in body and part header #2675
-
+* mailheader: strip whitespace between encoded-words #2702
 
 ## 2.8.24 - Mar 12, 2019